+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added documentation on how to use the logging library with `shamo`.
 - Added the `mesh_from_fem` method to provide a way to generate a mesh from a previously built one by merging tissues together.
 - Added the `mesh_from_surfaces` method to provide a way to generate a mesh from a set of surface meshes.
+- Added `SurrMaskedScalar` and `SurrMaskedScalarNii` to enable more sensitivity analysis.
+- Added `SurrMaskedScalarNiiJ`, `SurrMaskedScalarNiiMagJ` and `SurrMaskedScalarNiiV` for tDCS.
 
 ### Changed